requirements.txt
annotated-types==0.6.0
anyio==3.7.1
azure-core==1.29.5
azure-data-tables==12.4.4
azure-identity==1.15.0
azure-kusto-data==4.2.0
azure-kusto-ingest==4.2.0
azure-storage-blob==12.19.0
azure-storage-queue==12.8.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
distro==1.8.0
Flask==3.0.0
Flask-Cors==4.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
ijson==3.2.3
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msal==1.25.0
msal-extensions==1.0.0
multidict==6.0.4
openai==1.3.4
portalocker==2.8.2
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pywin32==306
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
tenacity==8.2.3
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.0
yarl==1.9.3
